requirements.txt
langtrace-python-sdk
crewai
crewai-tools
google_generativeai
python-dotenv